For SE Class

+++ src/main/java/com/amit/primesservices/services/IPrimeService.java
@@ -1,5 +1,5 @@
 package com.amit.primesservices.services;
 
 public interface IPrimeService {
-    boolean isPrime(int n);
+    boolean isPrime(long n);
 }
